Fix array count read on 32-bit targets
The number of records preceding a libyang array (LY_ARRAY_COUNT) is
stored as a uint64_t regardless of the pointer width. Reading it as a
`usize` yields a bogus count on targets where `usize` isn't 64 bits,
such as wasm32.

+++ src/iter.rs
@@ -347,11 +347,12 @@ where
         ptr_size: usize,
     ) -> Array<'a, S> {
         // Get the number of records in the array (equivalent to
-        // LY_ARRAY_COUNT).
+        // LY_ARRAY_COUNT). It is stored in the LY_ARRAY_COUNT_TYPE preceding
+        // the array, which is a uint64_t regardless of the pointer width.
         let count = if raw.is_null() {
             0
         } else {
-            unsafe { (raw as *const usize).offset(-1).read() }
+            unsafe { (raw as *const u64).offset(-1).read() as usize }
         };
 
         Array {
